CoRC: COPASI R Connector.
Source: R/CoRC.R

CoRC, the COPASI R Connector, links the Complex Pathway Simulator COPASI to R. It provides easy access to the powerful biochemical model editing, simulation and analysis backend of COPASI. This allows the user to develop elaborate scripts and workflows for analyses that would require a great deal of tedious manual work otherwise. These scripts can then be run interactively or be sent to cluster or cloud facilities for more demanding calculations.

Details

CoRC features:

* high-level API for COPASI in the R language. * Immediate access to R's data analysis capabilities and publication-ready plotting. * Reproducible workflows from data generation to analysis and plotting (R scripts and notebooks). * Rule-based modification of model structure to test structural variations or create large models. * Scaling up assays, e.g. from 3 models to 3000. * Handling of multiple models at once. * Parallelization on multi-core machines or computing clusters.

It is based on a current development version of COPASI but is expected to closely follow official COPASI releases in the future.

See also