Define a parameter estimation experiment

defineExperiments defines a set of experiments given as tidy data frame to the given model.

defineExperiments(
  experiment_type = c("time_course", "steady_state"),
  data = NULL,
  types = NULL,
  mappings = NULL,
  weights = NULL,
  weight_method = NULL,
  normalize_weights_per_experiment = FALSE,
  filename = NULL
)

Arguments

experiment_type

string

data

list of tidy data frames

types

data column types as character vector

Allowed types of columns are: 'time', 'independent', 'dependent', 'ignore'.

Type 'time' is only allowed for time course experiments.

Can be a named vector to only specify for a subset of data columns.

mappings

data column mappings as character vector

Expects value references.

If no mappings are given, column names can serve as mappings.

Can be a named vector to only specify for a subset of data columns.

weights

data column weights as numeric vector

`NaN` corresponds to automatic weight.

Can be a named vector to only specify for a subset of data columns.

weight_method

string

Allowed methods: 'mean', 'mean_square', 'sd', 'value_scaling'.

normalize_weights_per_experiment

flag

filename

optional string

When adding the experiments to a COPASI model, this filename will be used. In use cases, where experiments are only used temporarily, the filename is ignored.

Value

copasi_exp object for input into related functions

Details

CoRC uses it's own methodology for defining experimental data for use with a COPASI model. To this end it is required that experimental data be imported to R by the user and transformed to tidy data. For help on data import and tidying see: vignette("tidy-data", "tidyr"). This function adds required metadata to experimental data for use with CoRC.

See also