Create a new kinetic function

The online article on model building provides some further context.

newKineticFunction(name, formula, parameters, function_type = c("general",
  "reversible", "irreversible"))

Arguments

name
Name to set, as string

formula
Function formula to set, as string

parameters
Parameter types to set, as named character vector. Defaults to 'parameter' for every parameter.

Allowed values: 'substrate', 'product', 'modifier', 'parameter', 'volume', 'time'.

function_type
Type of function to set, as string. Defines if the function is general, reversible or irreversible.

Value

function key

See also