Run optimization
Source: R/task_optimization.R

runOptimization runs optimization and returns the results in a list.

runOptimization(expression = NULL, maximize = NULL, subtask = NULL,
  randomize_start_values = NULL, calculate_statistics = NULL,
  update_model = NULL, executable = NULL, parameters = NULL,
  method = NULL, model = getCurrentModel())

runOpt(expression = NULL, maximize = NULL, subtask = NULL,
  randomize_start_values = NULL, calculate_statistics = NULL,
  update_model = NULL, executable = NULL, parameters = NULL,
  method = NULL, model = getCurrentModel())

Arguments

expression
Expression to optimize, as string.

maximize
flag

subtask
string

Available tasks: 'steadyState', 'timeCourse', 'scan', 'fluxMode', 'optimization', 'parameterFitting', 'mca', 'lyap', 'tssAnalysis', 'sens', 'moieties', 'crosssection', 'lna', 'analytics'.

randomize_start_values
flag

calculate_statistics
flag

update_model
flag

executable
flag

parameters
copasi_param or list of copasi_param objects

method
string or list

Set the method directly with a string or with a list containing the entry named `method`.

The list may contain further method arguments and their values.

Available methods: 'Statistics', 'DifferentialEvolution', 'SRES', 'EvolutionaryProgram', 'GeneticAlgorithm', 'GeneticAlgorithmSR', 'HookeJeeves', 'LevenbergMarquardt', 'NelderMead', 'ParticleSwarm', 'Praxis', 'RandomSearch', 'ScatterSearch', 'SimulatedAnnealing', 'SteepestDescent', 'TruncatedNewton'.

model
A model object.

Value

A list of results.

See also