Run parameter estimation

runParameterEstimation runs parameter estimation and returns the results in a list.

runParameterEstimation(randomize_start_values = NULL,
  create_parameter_sets = NULL, calculate_statistics = NULL,
  update_model = NULL, executable = NULL, parameters = NULL,
  experiments = NULL, method = NULL, model = getCurrentModel())

runPE(randomize_start_values = NULL, create_parameter_sets = NULL,
  calculate_statistics = NULL, update_model = NULL, executable = NULL,
  parameters = NULL, experiments = NULL, method = NULL,
  model = getCurrentModel())

Arguments

randomize_start_values
flag

create_parameter_sets
flag

calculate_statistics
flag

update_model
flag

executable
flag

parameters
copasi_param or list of copasi_param objects

experiments
copasi_exp or list of copasi_exp objects

method
string or list

Set the method directly with a string or with a list containing the entry named `method`.

The list may contain further method args and their values.

Allowed methods: 'Statistics', 'DifferentialEvolution', 'SRES', 'EvolutionaryProgram', 'GeneticAlgorithm', 'GeneticAlgorithmSR', 'HookeJeeves', 'LevenbergMarquardt', 'NelderMead', 'ParticleSwarm', 'Praxis', 'RandomSearch', 'ScatterSearch', 'SimulatedAnnealing', 'SteepestDescent', 'TruncatedNewton'.

model
A model object.

Value

A list of results.

See also

Other parameter estimation: addExperiments, addParameterEstimationParameter, clearExperiments, clearParameterEstimationParameters, clearValidations, defineExperiments, getParameterEstimationSettings, setParameterEstimationSettings