Run sigma point method

runSigmaPoint runs parallel parameter estimations as dictated by the sigma point method and returns the results in a list.

runSigmaPoint(
  alpha = 0.5,
  beta = 2,
  kappa = 3,
  var = NULL,
  experiments,
  mean_fit_as_basis = TRUE,
  cl = NULL,
  model = getCurrentModel()
)

runSP(
  alpha = 0.5,
  beta = 2,
  kappa = 3,
  var = NULL,
  experiments,
  mean_fit_as_basis = TRUE,
  cl = NULL,
  model = getCurrentModel()
)

Arguments

alpha

number

beta

number

kappa

count

var

Measurement variance, as numeric. Can be a scalar value for variance on all measurements. A vector for variance per dependent experimental column or all measurements individually. A matrix for full covariance matrix.

experiments

Object of type copasi_exp as returned by defineExperiments. If it contains a single experiment, the values are understood to be mean values. If it contains multiple experiments, the experiments have to have identical independent columns.

mean_fit_as_basis

Whether to initially fit to the mean experimental data and set those fitted parameter values as starting values. Given as flag.

cl

Either object of type 'cluster' as used in the parallel package, or a count of threads to use, or NULL to automatically use all cores on current machine.

model

A model object.

Value

A list of results.

$fit_task_results is a list of all parameter estimation results gathered in this assay.

$sp_means is the result of term (28).

$sp_cov_matrix is the result of term (29).

$param_bias is the result of term (33).

$param_cov_matrix is the result of term (34).

Details

WARNING: TODO: This implementation of the sigma point method is almost completely untested.

Uses the current parameter estimation settings as defined in the model, with the exception of experiments.

As described in:
Schenkendorf, R., Kremling, A., & Mangold, M. (2009).
Optimal experimental design with the sigma point method.
IET Systems Biology, 3(1), 10–23.
https://doi.org/10.1049/iet-syb:20080094