Run a time course

runTimeCourse runs a time course and returns the results in a list.

runTimeCourse(duration = NULL, dt = NULL, intervals = NULL,
  suppress_output_before = NULL, output_events = NULL,
  save_result_in_memory = NULL, start_in_steady_state = NULL,
  update_model = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())

runTC(duration = NULL, dt = NULL, intervals = NULL,
  suppress_output_before = NULL, output_events = NULL,
  save_result_in_memory = NULL, start_in_steady_state = NULL,
  update_model = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())

Arguments

duration
The time course duration, as number.

dt
The time course output step size, as number.

intervals
The time course step count, as count. Overwrites dt in case of conflict.

suppress_output_before
Whether to suppress before a certain time point, as number.

output_events
Whether to output events as additional steps, as flag.

save_result_in_memory
Whether to generate an output data frame for the time course, as flag.

start_in_steady_state
Whether to first go to steady state before running the time course, as flag.

update_model
Whether to update the model with the state resulting from the time course, as flag.

executable
flag

method
string or list

Set the method directly with a string or with a list containing the entry named `method`.

The list may contain further method arguments and their values.

Available methods: 'deterministic', 'stochastic', 'directMethod', 'tauLeap', 'adaptiveSA', 'hybrid', 'hybridLSODA', 'hybridODE45', 'stochasticRunkeKuttaRI5'.

model
A model object.

Value

A list of results.

$result is a data frame containing time course output as concentrations.

$result_number is a data frame containing time course output as particle numbers.

Attention: Both result data frames will be empty if save_result_in_memory is set to FALSE.

Details

The online article on managing tasks provides some further context.

See also