Entity finders

The entity finder family of functions are a set of helpers to get unique identifiers for the various entities defined in COPASI. They give a quick overview of available species, compartments, etc.

species(key = "", reference = NULL, model = getCurrentModel())

species_strict(key, reference = NULL, model = getCurrentModel())

quantity(key = "", reference = NULL, model = getCurrentModel())

quantity_strict(key, reference = NULL, model = getCurrentModel())

compartment(key = "", reference = NULL, model = getCurrentModel())

compartment_strict(key, reference = NULL, model = getCurrentModel())

reaction(key = "", reference = NULL, model = getCurrentModel())

reaction_strict(key, reference = NULL, model = getCurrentModel())

parameter(key = "", reference = NULL, model = getCurrentModel())

parameter_strict(key, reference = NULL, model = getCurrentModel())

event(key = "", model = getCurrentModel())

event_strict(key, model = getCurrentModel())

kinfunction(key = "")

kinfunction_strict(key)

Arguments

key

entity name fragments

* main varieties: a string.

* `_strict` varieties: a character vector with each entry uniquely matching one entity.

reference

an optional string naming the value references to be returned.

model

a model object.

Value

A character vector of species identifiers or references.

References are in the form of "{name}" or in rare cases in the form of "<CN=...>" and can be used in expressions or functions such as getValue.

Details

There are two classes of entity finders.

The main functions are flexible and return a vector of all entities matching a given name fragment.

The `_strict` varieties expect for all input keys to uniquely match one entity and throw errors otherwise. This ensures that there are always as many identifiers returned as are given to the functions. This mechanism is also used by CoRC internally to ensure that functions like getSpecies return one row in the output for each key given as parameter.

Whereas the main functions generally won't help separate an entity `a` from and entity `ab`, the `_strict` varieties will accept a key "a" because it fully matches the entity `a`, which gets preference over the partial match with `ab`.

The matching mechanism can be tuned using a regex mechanism.

The functions can also be used to find value references by giving the `reference` argument. Various references such as "Value" or "InitialConcentration" are available for some types of entities.

The online article on managing model entities provides some further context.

See also