Run a time course

runTimeCourse runs a time course and returns the results in a list.

runTimeCourse(
  duration = NULL,
  dt = NULL,
  intervals = NULL,
  automatic_intervals = NULL,
  suppress_output_before = NULL,
  output_events = NULL,
  save_result_in_memory = NULL,
  start_in_steady_state = NULL,
  update_model = NULL,
  executable = NULL,
  method = NULL,
  soft_error = FALSE,
  model = getCurrentModel()
)

runTC(
  duration = NULL,
  dt = NULL,
  intervals = NULL,
  automatic_intervals = NULL,
  suppress_output_before = NULL,
  output_events = NULL,
  save_result_in_memory = NULL,
  start_in_steady_state = NULL,
  update_model = NULL,
  executable = NULL,
  method = NULL,
  soft_error = FALSE,
  model = getCurrentModel()
)

Arguments

duration
The time course duration, as number.

dt
The time course output step size, as number.

intervals
The time course step count, as count. Overwrites dt in case of conflict.

automatic_intervals
Whether to use automatic intervals, as flag.

suppress_output_before
Whether to suppress before a certain time point, as number.

output_events
Whether to output events as additional steps, as flag.

save_result_in_memory
Whether to generate an output data frame for the time course, as flag.

start_in_steady_state
Whether to first go to steady state before running the time course, as flag.

update_model
Whether to update the model with the state resulting from the time course, as flag.

executable
flag

method
string or list

Set the method directly with a string or with a list containing the entry named `method`.

The list may contain further method arguments and their values.

Available methods: 'deterministic', 'RADAU5', 'stochastic', 'directMethod', 'tauLeap', 'adaptiveSA', 'hybrid', 'hybridLSODA', 'hybridODE45', 'stochasticRunkeKuttaRI5'.

soft_error
Whether to divert task processing errors to the task_error result entry, as flag. Allows for readout of incomplete result tables for failed time courses.

model
A model object.

Value

A list of results.

$result is a data frame containing time course output as concentrations.

$result_number is a data frame containing time course output as particle numbers.

$task_error is NULL or, in case soft_error is set to TRUE and the time course failed, a string giving an error message.

Attention: Both result data frames will be empty if save_result_in_memory is set to FALSE. Test the current value of this parameter with: getTimeCourseSettings()$save_result_in_memory.

Details

The online article on managing tasks provides some further context.

See also